fix(CustomEndpoints): created a util function for disabling the submit button and type fixes
