+++ apps/Conduit-UI/src/models/customEndpoints/customEndpointsModels.ts
@@ -1,3 +1,5 @@
+import { EndpointInputs } from '../database/CmsModels';
+
 export interface Assignment {
   schemaField: string;
   action: number;
@@ -13,4 +15,22 @@ export interface Input {
   type: string;
   optional?: boolean;
   array?: boolean;
+}
+
+export interface Endpoint {
+  _id: string;
+  name: string;
+  operation: any;
+  selectedSchema: string;
+  authentication: boolean;
+  paginated?: boolean;
+  sorted?: boolean;
+  inputs: EndpointInputs[];
+  queries: [];
+  assignments: Assignment[];
+  enabled?: boolean;
+  selectedSchemaName?: string;
+  returns?: string;
+  createdAt?: string;
+  updatedAt?: string;
 }
